requirement.txt
torch
torchvision
torchaudio
numpy
pandas
pillow
matplotlib
scikit-learn